Fixing user point model.

+++ src/main/java/com/changing/party/exception/UserIdNotFoundException.java
@@ -2,7 +2,7 @@ package com.changing.party.exception;
 
 public class UserIdNotFoundException extends RuntimeException {
 
-    public UserIdNotFoundException(Long userId) {
+    public UserIdNotFoundException(int userId) {
         super(String.format("User id %s not found", userId));
     }
 }
